Replace manifest.json with @now/frameworks
