+++ examples/create-react-app/README.md
@@ -1,20 +1,20 @@
 ![React Logo](../.github/images/react.svg)
 
-# Create-React-App Example
+# React Example
 
-This directory is a brief example of a [React](https://reactjs.org/) app that can be deployed with ZEIT Now and zero configuration.
+This directory is a brief example of a [React](https://reactjs.org/) app with [Serverless Functions](https://zeit.co/docs/v2/serverless-functions/introduction) that can be deployed with ZEIT Now and zero configuration.
 
 ## Deploy Your Own
 
-Deploy your own React project with ZEIT Now.
+Deploy your own React project, along with Serverless Functions, with ZEIT Now.
 
-[![Deploy with ZEIT Now](https://zeit.co/button)](https://zeit.co/new/project?template=https://github.com/zeit/now-examples/tree/master/create-react-app)
+[![Deploy with ZEIT Now](https://zeit.co/button)](https://zeit.co/new/project?template=https://github.com/zeit/now-examples/tree/master/create-react-app-functions)
 
-_Live Example: https://create-react-app.now-examples.now.sh_
+_Live Example: https://create-react-app-functions.now-examples.now.sh/_
 
 ### How We Created This Example
 
-To get started with React, you can use the [Create-React-App CLI](https://reactjs.org/docs/create-a-new-react-app.html#create-react-app) to initialize the project:
+To get started with React, along with [Serverless Functions](https://zeit.co/docs/v2/serverless-functions/introduction), with ZEIT Now, you can use the [Create-React-App CLI](https://reactjs.org/docs/create-a-new-react-app.html#create-react-app) to initialize the project:
 
 ```shell
 $ npx create-react-app my-app
@@ -22,7 +22,7 @@ $ npx create-react-app my-app
 
 ### Deploying From Your Terminal
 
-You can deploy your new React project with a single command from your terminal using [Now CLI](https://zeit.co/download):
+You can deploy your new React project, along with [Serverless Functions](https://zeit.co/docs/v2/serverless-functions/introduction), with a single command from your terminal using [Now CLI](https://zeit.co/download):
 
 ```shell
 $ now
