+++ examples/create-react-app/api/date.go
