--- examples/umi/.env
@@ -1,2 +1,0 @@
-BROWSER=none
-ESLINT=1
