--- examples/umi/.nowignore
@@ -1,2 +1,0 @@
-README.md
-yarn.lock
