+++ examples/umijs/src/assets/yay.jpg
