+++ examples/umijs/tsconfig.json
