+++ packages/frameworks/index.d.ts
@@ -3,12 +3,13 @@ export interface FrameworkDetectionItem {
   matchContent?: string;
 }
 
-type Setting = { value: string } | { placeholder: string }
+type Setting = { value: string } | { placeholder: string };
 
 export interface Framework {
   name: string;
   slug: string;
   logo: string;
+  demo?: string;
   tagline: string;
   website: string;
   detectors?: {
